Split unit tests into two applications
So far, the two applications are indentical, except for their linkage
(one is to the dynamaic library, the other to the library source code
itself).  Eventually, the "api" version will test just the publically
exposed API (ie it could be shipped, and executed against other ABI-
compatible libraries), while the "all" version will continue to test
all of the library's functions - including the private internals, and
as such, would not work with any other version of the library.

This will also help significantly with resolving issue #9.