Disable copying of all objects with D-pointers
Except for the SqsError class, for which I've specifically implemented
the copy constructor and copy assignment operator for QList
compatibility.

Note, it might make sense to do this for the Sqs*Request classes too,
but I'd prefer them to be copyable, and they don't currently have (nor
need) D-pointers.  That will require some more thinking.