Don't build docs/tests if the lib failed to build
This should fix the occassional Travis CI build failure that appears to
result from make being invoked with `-j2` allowing the `test/unit/api`
project to be built before the library has finished building.

This change is applied to the other test projects, and the doxygen
output too, because I'd rather not waste time building those if/when
the library fails to build.