Remove a compiler-optimisation sensitvity
Specifically, gcc was optimising the recursive `::toVariant` invocation
in such as way that the recursive call was being performed before the
current XML element's name was being fetched.  This, of course, is
perfectly valid for C/C++ compilers to do.

The fix is simply to fetch the XML element's name explicitly before the
recursion.