Make the request classes copyable
Still need to doucment and unit-test the new functions.  Also still
need to add copy constructors and assignment operators to the not-yet-
implemented SQS request classes, but they can be done at the time of
implementation.

This, of course, if all preparatory work for allowing *Response objects
to maintain a copy of their respective *Request objects.  That will
come soon (next?).