Add the initial auto-generated client versions!
How exciting!!  These are incomplete, and are in particular are missing
requisite operations and model code, but it's a start! :)

Note, this code is not yet included in the higher-level qmake project
files.  They will not compile cleanly yet, so no put adding them yet.
But by including the incomplete auto-generated classes now, we can
more easily track changes to the auto-generator as they are made.